mixin fixes & float sprites

+++ common/src/main/java/net/mehvahdjukaar/slotify/SimpleSprite.java
@@ -12,16 +12,16 @@ import org.joml.Matrix4f;
 
 import java.util.Optional;
 
-public record SimpleSprite(ResourceLocation texture, int x, int y, int width, int height, int z,
+public record SimpleSprite(ResourceLocation texture, float x, float y, float width, float height, float z,
                            Optional<String> tooltip){//, Optional<ScreenSupplier> screenSupp) {
 
     public static final Codec<SimpleSprite> CODEC = RecordCodecBuilder.create(i -> i.group(
             ResourceLocation.CODEC.fieldOf("texture").forGetter(SimpleSprite::texture),
-            Codec.INT.fieldOf("x").forGetter(SimpleSprite::x),
-            Codec.INT.fieldOf("y").forGetter(SimpleSprite::y),
-            Codec.INT.fieldOf("width").forGetter(SimpleSprite::width),
-            Codec.INT.fieldOf("height").forGetter(SimpleSprite::height),
-            Codec.INT.optionalFieldOf("z_offset", 0).forGetter(SimpleSprite::z),
+            Codec.FLOAT.fieldOf("x").forGetter(SimpleSprite::x),
+            Codec.FLOAT.fieldOf("y").forGetter(SimpleSprite::y),
+            Codec.FLOAT.fieldOf("width").forGetter(SimpleSprite::width),
+            Codec.FLOAT.fieldOf("height").forGetter(SimpleSprite::height),
+            Codec.FLOAT.optionalFieldOf("z_offset", 0.0f).forGetter(SimpleSprite::z),
             Codec.STRING.optionalFieldOf("tooltip").forGetter(SimpleSprite::tooltip)
            // Codec.STRING.xmap(ScreenSupplier::decode, ScreenSupplier::toString).f
                  //   .optionalFieldOf("screen_class").forGetter(SimpleSprite:: screenSupp)
@@ -34,14 +34,14 @@ public record SimpleSprite(ResourceLocation texture, int x, int y, int width, in
         innerBlit(poseStack.last().pose(), x, x+width, y, y+height, z);
     }
 
-    private static void innerBlit(Matrix4f matrix, int x1, int x2, int y1, int y2, int blitOffset) {
+    private static void innerBlit(Matrix4f matrix, float x1, float x2, float y1, float y2, float blitOffset) {
         RenderSystem.setShader(GameRenderer::getPositionTexShader);
         BufferBuilder bufferBuilder = Tesselator.getInstance().getBuilder();
         bufferBuilder.begin(VertexFormat.Mode.QUADS, DefaultVertexFormat.POSITION_TEX);
-        bufferBuilder.vertex(matrix, (float)x1, (float)y2, (float)blitOffset).uv(0, 1).endVertex();
-        bufferBuilder.vertex(matrix, (float)x2, (float)y2, (float)blitOffset).uv(1, 1).endVertex();
-        bufferBuilder.vertex(matrix, (float)x2, (float)y1, (float)blitOffset).uv(1, 0).endVertex();
-        bufferBuilder.vertex(matrix, (float)x1, (float)y1, (float)blitOffset).uv(0, 0).endVertex();
+        bufferBuilder.vertex(matrix, x1, y2, blitOffset).uv(0, 1).endVertex();
+        bufferBuilder.vertex(matrix, x2, y2, blitOffset).uv(1, 1).endVertex();
+        bufferBuilder.vertex(matrix, x2, y1, blitOffset).uv(1, 0).endVertex();
+        bufferBuilder.vertex(matrix, x1, y1, blitOffset).uv(0, 0).endVertex();
         BufferUploader.drawWithShader(bufferBuilder.end());
     }
 }
